Fix AMR,RAAC,COOK 64bit play error

+++ audio_codec/libamr/typedef.h
@@ -10,7 +10,7 @@ typedef char Word8;
 typedef unsigned char UWord8;
 typedef short Word16;
 typedef unsigned short UWord16;
-typedef long Word32;
+typedef int Word32;
 typedef double Float64;
 typedef float Float32;
